refactor(userRoom): userRoomRepository로 메서드 분리
